--- packages/client/src/pages/UpdateProfilePage.tsx
@@ -1,156 +1,0 @@
-import { useState } from 'react'
-import type { FormEvent } from 'react'
-import { useAuth } from '../auth/AuthContext'
-import { authApi } from '../auth/api'
-
-const UpdateProfilePage = () => {
-  const { user } = useAuth()
-  const [name, setName] = useState(user?.name || '')
-  const [password, setPassword] = useState('')
-  const [confirmPassword, setConfirmPassword] = useState('')
-  const [loading, setLoading] = useState(false)
-  const [error, setError] = useState('')
-  const [success, setSuccess] = useState('')
-
-  const handleSubmit = async (e: FormEvent) => {
-    e.preventDefault()
-    setError('')
-    setSuccess('')
-
-    // Validate passwords match
-    if (password && password !== confirmPassword) {
-      setError('Passwords do not match')
-      return
-    }
-
-    setLoading(true)
-
-    try {
-      // Only send fields that have been changed
-      const updateData: any = {}
-      if (name && name !== user?.name) {
-        updateData.name = name
-      }
-      if (password) {
-        updateData.password = password
-      }
-
-      // Only make API call if there are changes
-      if (Object.keys(updateData).length > 0) {
-        await authApi.updateProfile(updateData.name, updateData.password)
-        setSuccess('Profile updated successfully')
-        setPassword('')
-        setConfirmPassword('')
-      } else {
-        setError('No changes to update')
-      }
-    } catch (err: any) {
-      setError(err.message || 'Failed to update profile')
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  return (
-    <div style={{ maxWidth: '400px', margin: '50px auto', padding: '20px' }}>
-      <h1>Update Profile</h1>
-      <form onSubmit={handleSubmit}>
-        <div style={{ marginBottom: '15px' }}>
-          <label style={{ display: 'block', marginBottom: '5px' }}>
-            Email (cannot be changed)
-          </label>
-          <input
-            type="email"
-            value={user?.email || ''}
-            disabled
-            style={{ width: '100%', padding: '8px', backgroundColor: '#f5f5f5' }}
-          />
-        </div>
-
-        <div style={{ marginBottom: '15px' }}>
-          <label style={{ display: 'block', marginBottom: '5px' }}>
-            Role (cannot be changed)
-          </label>
-          <input
-            type="text"
-            value={user?.role || ''}
-            disabled
-            style={{ width: '100%', padding: '8px', backgroundColor: '#f5f5f5' }}
-          />
-        </div>
-
-        <div style={{ marginBottom: '15px' }}>
-          <label htmlFor="name" style={{ display: 'block', marginBottom: '5px' }}>
-            Name
-          </label>
-          <input
-            type="text"
-            id="name"
-            value={name}
-            onChange={(e) => setName(e.target.value)}
-            required
-            style={{ width: '100%', padding: '8px' }}
-          />
-        </div>
-
-        <div style={{ marginBottom: '15px' }}>
-          <label htmlFor="password" style={{ display: 'block', marginBottom: '5px' }}>
-            New Password (leave empty to keep current)
-          </label>
-          <input
-            type="password"
-            id="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            minLength={8}
-            style={{ width: '100%', padding: '8px' }}
-          />
-        </div>
-
-        <div style={{ marginBottom: '15px' }}>
-          <label htmlFor="confirmPassword" style={{ display: 'block', marginBottom: '5px' }}>
-            Confirm New Password
-          </label>
-          <input
-            type="password"
-            id="confirmPassword"
-            value={confirmPassword}
-            onChange={(e) => setConfirmPassword(e.target.value)}
-            minLength={8}
-            style={{ width: '100%', padding: '8px' }}
-          />
-        </div>
-
-        {error && (
-          <div style={{ color: 'red', marginBottom: '15px' }}>
-            {error}
-          </div>
-        )}
-
-        {success && (
-          <div style={{ color: 'green', marginBottom: '15px' }}>
-            {success}
-          </div>
-        )}
-
-        <button
-          type="submit"
-          disabled={loading}
-          style={{
-            width: '100%',
-            padding: '10px',
-            backgroundColor: '#007bff',
-            color: 'white',
-            border: 'none',
-            cursor: loading ? 'not-allowed' : 'pointer',
-            opacity: loading ? 0.7 : 1,
-          }}
-        >
-          {loading ? 'Updating...' : 'Update Profile'}
-        </button>
-      </form>
-    </div>
-  )
-}
-
-export default UpdateProfilePage
